# api-spec/openapi/pilot-v1.yaml
...
openapi: 3.1.0
...
info:
  title: Pilot API
  version: 1.0.0
  contact:
    name: bcodesphere
  description: >-
    API REST de Pilot, ERP multiempresa para PYMES de El Salvador. Incluye el núcleo
    (usuarios y empresas) y la app Contabilidad. Los montos viajan como cadena decimal
    con 2 decimales (ADR-013) y los errores siguen RFC 9457 (Problem Details).

# --- Servidores: ruta base relativa al host que sirve la API ---
servers:
  - url: /api/v1
    description: Ruta base de la versión 1 de la API.

# --- Etiquetas por módulo del backend (CLAUDE.md §4.2) ---
...
security:
  - oidc: []
paths:
...
  /contabilidad/asientos:
    post:
      operationId: registrarAsiento
      tags: [asientos]
      x-rol-minimo: contador
      summary: Registra un asiento manual
      description: >-
        Registra un asiento manual del Libro Diario en una sola transacción: expande el IVA de las líneas que lo llevan
        (CLAUDE.md §11.2), valida la partida doble, asigna el número correlativo, guarda cabecera y líneas y mayoriza
        los saldos (ADR-018); o se guarda todo o no se guarda nada. Exige Idempotency-Key: repetir la misma clave con el
        mismo cuerpo devuelve la respuesta original con Idempotency-Replayed true. Un asiento guardado es inmutable
        (ADR-019). Rol mínimo: contador.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        description: Fecha, concepto, modo de precio opcional y líneas del asiento.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevoAsiento'
      responses:
        '201':
          description: Asiento contabilizado. Si es la repetición de una petición idempotente, es la respuesta original.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            Idempotency-Replayed:
              $ref: '#/components/headers/Idempotency-Replayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asiento'
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: cuerpo JSON ilegible o con tipos incorrectos (incluye un monto enviado como número)."
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '409':
          $ref: '#/components/responses/Conflicto'
          description: "409 — PLT-008: otra petición con la misma Idempotency-Key está en proceso; reintentar."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — CON-001: menos de dos líneas; CON-002: una línea con Debe y Haber a la vez o ninguno; CON-003: monto negativo o con más de 2 decimales; CON-004: totales en cero; CON-005: Σ Debe ≠ Σ Haber (incluye diferencia); CON-006: cuenta inexistente, inactiva, de otra empresa o que no es de detalle; CON-007: fecha obligatoria y no futura; CON-013: \"lleva IVA\" sobre una cuenta de IVA; CON-017: no hay tasa de IVA vigente a la fecha del asiento; PLT-002: validación de forma (concepto obligatorio o de más de 500 caracteres, más de 200 líneas; incluye errores); PLT-005: Idempotency-Key reutilizada con otro cuerpo."
        '428':
          $ref: '#/components/responses/PrecondicionRequerida'
          description: "428 — PLT-006: falta el header Idempotency-Key."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
    get:
      operationId: listarAsientos
      tags: [asientos]
      x-rol-minimo: auditor
      summary: Libro Diario paginado
      description: >-
        Lista los asientos de la empresa (Libro Diario) paginados por cursor y ordenados por año y número ascendentes.
        Los filtros son opcionales y se combinan: desde y hasta (fecha contable, ambos inclusive), anio, numero, origen,
        estado y cuentaId (asientos con alguna línea en esa cuenta). El listado devuelve el resumen sin líneas; el detalle
        está en la consulta individual. Rol mínimo: auditor.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/Limite'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Desde'
        - $ref: '#/components/parameters/Hasta'
        - $ref: '#/components/parameters/AnioFiltro'
        - $ref: '#/components/parameters/NumeroAsientoFiltro'
        - $ref: '#/components/parameters/OrigenAsientoFiltro'
        - $ref: '#/components/parameters/EstadoAsientoFiltro'
        - $ref: '#/components/parameters/CuentaIdFiltro'
      responses:
        '200':
          description: Página de asientos en orden de año y número.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaAsientos'
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: parámetros con tipos incorrectos."
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — PLT-002: validación de forma de los parámetros, o desde es posterior a hasta (incluye errores)."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /contabilidad/asientos/vista-previa:
    post:
      operationId: previsualizarAsiento
      tags: [asientos]
      x-rol-minimo: contador
      summary: Vista previa de un asiento sin guardarlo
      description: >-
        Expande el IVA de las líneas que lo llevan y calcula los totales y la diferencia sin guardar nada ni numerar el
        asiento; no exige Idempotency-Key. Responde 200 aunque el asiento no cuadre (cuadra false y diferencia distinta de
        cero, ADR-036), para que el frontend valide la partida doble sobre las líneas expandidas por el backend. Solo
        responde 422 si no se puede expandir. Rol mínimo: contador.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
      requestBody:
        required: true
        description: Mismo cuerpo que el registro del asiento.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevoAsiento'
      responses:
        '200':
          description: Líneas expandidas, totales, diferencia e indicador de que cuadra.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VistaPreviaAsiento'
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
          description: "400 — PLT-001: cuerpo JSON ilegible o con tipos incorrectos (incluye un monto enviado como número)."
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '422':
          $ref: '#/components/responses/EntidadNoProcesable'
          description: "422 — solo si no se puede expandir: CON-002: una línea con Debe y Haber a la vez o ninguno; CON-003: monto negativo o con más de 2 decimales; CON-006: cuenta inexistente, inactiva, de otra empresa o que no es de detalle; CON-007: fecha obligatoria y no futura; CON-013: \"lleva IVA\" sobre una cuenta de IVA; CON-017: no hay tasa de IVA vigente a la fecha; PLT-002: validación de forma (incluye errores)."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /contabilidad/asientos/{asientoId}:
    get:
      operationId: obtenerAsiento
      tags: [asientos]
      x-rol-minimo: auditor
      summary: Consulta un asiento con sus líneas
      description: >-
        Devuelve un asiento con todas sus líneas. No lleva ETag porque un asiento guardado es inmutable salvo su paso a
        REVERTIDO (ADR-019). Rol mínimo: auditor.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XEmpresaId'
        - $ref: '#/components/parameters/AsientoIdRuta'
      responses:
        '200':
          description: Asiento con sus líneas.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asiento'
        '401':
          $ref: '#/components/responses/NoAutenticado'
          description: "401 — PLT-009: falta el token, es inválido o está vencido."
        '403':
          $ref: '#/components/responses/Prohibido'
          description: "403 — PLT-003: el usuario no tiene membresía activa en X-Empresa-Id; PLT-004: la app Contabilidad no está instalada en la empresa; PLT-010: el rol no alcanza para esta operación."
        '404':
          $ref: '#/components/responses/NoEncontrado'
          description: "404 — PLT-017: el recurso no existe o no pertenece a la empresa activa."
        '429':
          $ref: '#/components/responses/DemasiadasPeticiones'
        '500':
          $ref: '#/components/responses/ErrorInterno'
...
components:
...
  securitySchemes:
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso JWT emitido por Keycloak (OIDC con PKCE) para usuarios de la web.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: pk_xxxx.secreto
      description: API key de integración con formato pk_xxxx.secreto, usada por n8n; la empresa se deriva de la clave.

  # --- Esquemas reutilizables ---
  schemas:
    Monto:
      type: string
      pattern: '^\d{1,17}(\.\d{1,2})?$'
      examples: ['123.45']
      description: >-
        Monto en USD no negativo como cadena decimal con máximo 2 decimales (ADR-013); nunca número JSON.
        El servidor siempre responde con exactamente 2 decimales.
    MontoConSigno:
      type: string
      pattern: '^-?\d{1,17}(\.\d{1,2})?$'
      examples: ['-10.50']
      description: >-
        Monto en USD que puede ser negativo (p. ej. una diferencia), como cadena decimal con máximo
        2 decimales (ADR-013). El servidor siempre responde con exactamente 2 decimales.
    Uuid:
      type: string
      format: uuid
      description: Identificador UUID (versión 7 generado por la aplicación, ADR-010).
    Fecha:
      type: string
      format: date
      description: Fecha calendario ISO-8601 (AAAA-MM-DD); las fechas contables son de hora de El Salvador.
    FechaHora:
      type: string
      format: date-time
      description: Instante ISO-8601 en UTC.
...
    ProblemDetails:
      type: object
      description: Error según RFC 9457 (application/problem+json) con extensiones de Pilot.
      required: [type, title, status, codigo]
      properties:
        type:
          type: string
          description: URI que identifica el tipo de problema; about:blank si no hay uno específico.
        title:
          type: string
          description: Resumen corto del problema, en español.
        status:
          type: integer
          description: Código de estado HTTP de la respuesta.
        detail:
          type: string
          description: Explicación del problema para esta ocurrencia; sin trazas ni datos internos.
        instance:
          type: string
          description: URI o identificador de la ocurrencia concreta del problema.
        codigo:
          type: string
          pattern: '^(PLT|CON|INT)-\d{3}$'
          description: Código de negocio con prefijo del módulo (PLT plataforma, CON contabilidad, INT integración).
        errores:
          type: array
          description: Errores de validación por campo; solo en fallos de validación.
          items:
            $ref: '#/components/schemas/ErrorCampo'
        diferencia:
          $ref: '#/components/schemas/MontoConSigno'
          description: Diferencia Σ Debe − Σ Haber; solo en descuadres y cobros que no cuadran.
        operacionId:
          $ref: '#/components/schemas/Uuid'
          description: Operación externa ya contabilizada; solo en 409 INT-004 (ADR-039 §5).
        asientoId:
          $ref: '#/components/schemas/Uuid'
          description: Asiento generado por la operación ya contabilizada; solo en 409 INT-004 (ADR-039 §5).
    PaginaCursor:
      type: object
      description: Metadatos de paginación por cursor; los listados lo extienden con su arreglo de elementos.
      properties:
        siguienteCursor:
          type: [string, 'null']
          description: Cursor opaco para pedir la página siguiente; nulo cuando no hay más resultados.
    # --- Enumeraciones con nombre (se reutilizan en lugar de repetirse en línea) ---
...
    ModoPrecio:
      type: string
      enum: [CON_IVA, SIN_IVA]
      description: >-
        Cómo se interpreta un monto: CON_IVA (el monto incluye IVA) o SIN_IVA (el IVA se suma). Solo cambia la
        interpretación, no las tasas (ADR-015).
...
    ResumenCuenta:
      type: object
      description: Referencia corta a una cuenta contable, para mostrarla dentro de otros recursos.
      required: [id, codigo, nombre]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        codigo:
          type: string
          description: Código de la cuenta.
        nombre:
          type: string
          description: Nombre de la cuenta.
...
    EstadoAsiento:
      type: string
      enum: [CONTABILIZADO, REVERTIDO]
      description: Estado del asiento; REVERTIDO es el único cambio permitido tras guardarlo (ADR-019).
    OrigenAsiento:
      type: string
      enum: [MANUAL, N8N, REVERSION]
      description: Cómo nació el asiento — registro manual, operación recibida por n8n o reversión de otro asiento.
...
    NuevaLineaAsiento:
      type: object
      description: Línea de entrada de un asiento; cada línea lleva solo Debe o solo Haber (CON-002).
      required: [cuentaId, debe, haber]
      properties:
        cuentaId:
          $ref: '#/components/schemas/Uuid'
          description: Cuenta de detalle, activa y de la empresa (CON-006).
        descripcion:
          type: string
          maxLength: 300
          description: Descripción opcional de la línea.
        debe:
          $ref: '#/components/schemas/MontoEntrada'
          description: Monto al Debe; "0" o "0.00" si la línea va al Haber.
        haber:
          $ref: '#/components/schemas/MontoEntrada'
          description: Monto al Haber; "0" o "0.00" si la línea va al Debe.
        llevaIva:
          type: boolean
          default: false
          description: >-
            Si la línea lleva IVA; el backend la expande en la línea base y una línea de IVA del mismo lado (débito fiscal
            al Haber, crédito fiscal al Debe). No se permite sobre cuentas de IVA (CON-013).
    NuevoAsiento:
      type: object
      description: >-
        Datos de un asiento manual. No lleva mínimo de líneas en la forma para que el negocio responda CON-001, ni
        totales: los calcula el backend.
      required: [fecha, concepto, lineas]
      properties:
        fecha:
          $ref: '#/components/schemas/Fecha'
          description: Fecha contable en hora de El Salvador; obligatoria y no futura (CON-007).
        concepto:
          type: string
          minLength: 1
          maxLength: 500
          description: Concepto o glosa del asiento.
        modoPrecio:
          $ref: '#/components/schemas/ModoPrecio'
          description: Modo de precio para separar el IVA; si falta se usa el de la configuración contable (ADR-015).
        lineas:
          type: array
          maxItems: 200
          description: Líneas del asiento, máximo 200; el mínimo de dos (CON-001) lo valida el negocio.
          items:
            $ref: '#/components/schemas/NuevaLineaAsiento'
    LineaAsiento:
      type: object
      description: Línea de un asiento guardado, ya expandida.
      required: [id, numeroLinea, cuenta, descripcion, debe, haber, origenLinea, lineaBaseId]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        numeroLinea:
          type: integer
          minimum: 1
          description: Posición de la línea dentro del asiento, desde 1.
        cuenta:
          $ref: '#/components/schemas/ResumenCuenta'
        descripcion:
          type: [string, 'null']
          description: Descripción de la línea; nula si no tiene.
        debe:
          $ref: '#/components/schemas/Monto'
        haber:
          $ref: '#/components/schemas/Monto'
        origenLinea:
          $ref: '#/components/schemas/OrigenLinea'
        lineaBaseId:
          type: [string, 'null']
          format: uuid
          description: En una línea de IVA calculado, la línea que la originó; nulo en las demás.
    Asiento:
      type: object
      description: Asiento guardado con sus líneas; inmutable salvo su paso a REVERTIDO (ADR-019).
      required: [id, anio, numero, fecha, concepto, estado, origenTipo, origenId, modoPrecio, asientoRevertidoId, asientoReversionId, totalDebe, totalHaber, creadoEn, lineas]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        anio:
          type: integer
          description: Año de la fecha contable; junto con numero forma el correlativo.
        numero:
          type: integer
          format: int64
          description: Número correlativo del asiento dentro de la empresa y el año.
        fecha:
          $ref: '#/components/schemas/Fecha'
        concepto:
          type: string
          description: Concepto o glosa del asiento.
        estado:
          $ref: '#/components/schemas/EstadoAsiento'
        origenTipo:
          $ref: '#/components/schemas/OrigenAsiento'
        origenId:
          type: [string, 'null']
          format: uuid
          description: Operación externa que originó el asiento; nulo si no viene de n8n.
        modoPrecio:
          oneOf:
            - $ref: '#/components/schemas/ModoPrecio'
            - type: 'null'
          description: Modo de precio usado para separar el IVA; nulo si ninguna línea llevó IVA.
        asientoRevertidoId:
          type: [string, 'null']
          format: uuid
          description: En una reversión, el asiento que revierte; nulo en las demás.
        asientoReversionId:
          type: [string, 'null']
          format: uuid
          description: En un asiento revertido, su reversión; nulo si no está revertido.
        totalDebe:
          $ref: '#/components/schemas/Monto'
        totalHaber:
          $ref: '#/components/schemas/Monto'
        creadoEn:
          $ref: '#/components/schemas/FechaHora'
        lineas:
          type: array
          description: Líneas del asiento ordenadas por numeroLinea.
          items:
            $ref: '#/components/schemas/LineaAsiento'
...
    PaginaAsientos:
      description: Página del Libro Diario.
      allOf:
        - $ref: '#/components/schemas/PaginaCursor'
        - type: object
          required: [elementos]
          properties:
            elementos:
              type: array
              description: Asientos de la página.
              items:
                $ref: '#/components/schemas/ResumenAsiento'
...
    VistaPreviaAsiento:
      type: object
      description: >-
        Resultado de expandir y sumar un asiento sin guardarlo. Se devuelve aunque no cuadre (ADR-036), para validar
        la partida doble sobre las líneas expandidas por el backend.
      required: [modoPrecio, tasaIva, lineas, totalDebe, totalHaber, diferencia, cuadra]
      properties:
        modoPrecio:
          oneOf:
            - $ref: '#/components/schemas/ModoPrecio'
            - type: 'null'
          description: Modo de precio aplicado; nulo si ninguna línea lleva IVA.
        tasaIva:
          type: [string, 'null']
          pattern: '^\d{1,3}(\.\d{1,4})?$'
          examples: ['0.1300']
          description: Tasa de IVA vigente a la fecha del asiento como cadena decimal; nula si ninguna línea lleva IVA.
        lineas:
          type: array
          description: Líneas expandidas, con las de IVA calculado incluidas.
          items:
            $ref: '#/components/schemas/LineaVistaPrevia'
        totalDebe:
          $ref: '#/components/schemas/Monto'
        totalHaber:
          $ref: '#/components/schemas/Monto'
        diferencia:
          $ref: '#/components/schemas/MontoConSigno'
          description: Σ Debe − Σ Haber de las líneas expandidas.
        cuadra:
          type: boolean
          description: true si hay al menos dos líneas, los totales son mayores que cero y el Debe es igual al Haber.
...
  parameters:
    XEmpresaId:
      name: X-Empresa-Id
      in: header
      required: true
      description: Empresa activa de la sesión; se valida contra las membresías del usuario (CLAUDE.md §4.5).
      schema:
        $ref: '#/components/schemas/Uuid'
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador de correlación; se genera si no viene y se devuelve en la respuesta.
      schema:
        type: string
        maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Clave de idempotencia de la operación; repetir la misma clave con el mismo cuerpo devuelve la respuesta original.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    Limite:
      name: limite
      in: query
      required: false
      description: Cantidad máxima de elementos por página.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor opaco devuelto como siguienteCursor en la página anterior.
      schema:
        type: string
...
    AsientoIdRuta:
      name: asientoId
      in: path
      required: true
      description: Identificador del asiento.
      schema:
        $ref: '#/components/schemas/Uuid'
    Desde:
      name: desde
      in: query
      required: false
      description: Fecha contable inicial, inclusive; no puede ser posterior a hasta.
      schema:
        $ref: '#/components/schemas/Fecha'
    Hasta:
      name: hasta
      in: query
      required: false
      description: Fecha contable final, inclusive.
      schema:
        $ref: '#/components/schemas/Fecha'
    AnioFiltro:
      name: anio
      in: query
      required: false
      description: Año de la fecha contable de los asientos.
      schema:
        type: integer
        minimum: 2000
        maximum: 9999
    NumeroAsientoFiltro:
      name: numero
      in: query
      required: false
      description: Número correlativo del asiento (se combina con anio para ubicar uno solo).
      schema:
        type: integer
        format: int64
        minimum: 1
    OrigenAsientoFiltro:
      name: origen
      in: query
      required: false
      description: Limita el listado a un origen de asiento.
      schema:
        $ref: '#/components/schemas/OrigenAsiento'
    EstadoAsientoFiltro:
      name: estado
      in: query
      required: false
      description: Limita el listado a un estado de asiento.
      schema:
        $ref: '#/components/schemas/EstadoAsiento'
    CuentaIdFiltro:
      name: cuentaId
      in: query
      required: false
      description: Limita el listado a los asientos con alguna línea en esta cuenta.
      schema:
        $ref: '#/components/schemas/Uuid'

    # --- Operaciones guiadas y tablero (ADR-041) ---
...
  headers:
    X-Request-Id:
      description: Identificador de correlación de la petición, para soporte y trazas.
      schema:
        type: string
...
    Idempotency-Replayed:
      description: Vale true cuando la respuesta es la guardada de una petición idempotente repetida.
      schema:
        type: boolean
    Retry-After:
      description: Segundos que el cliente debe esperar antes de reintentar.
      schema:
        type: integer
...
  responses:
    SolicitudInvalida:
      description: 400 — El cuerpo o los parámetros no cumplen el contrato (incluye la lista errores).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NoAutenticado:
      description: 401 — Falta la credencial o es inválida, está vencida o fue revocada.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Prohibido:
      description: 403 — La credencial es válida pero no tiene rol, alcance o app activa para esta operación.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NoEncontrado:
      description: 404 — El recurso no existe o no pertenece a la empresa activa.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflicto:
      description: 409 — El estado actual del recurso impide la operación (duplicado o petición en proceso).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    EntidadNoProcesable:
      description: 422 — La petición es válida en forma pero viola una regla de negocio (incluye diferencia si aplica).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    PrecondicionRequerida:
      description: 428 — Falta un header obligatorio de precondición, como Idempotency-Key o If-Match.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    DemasiadasPeticiones:
      description: 429 — Se superó el límite de peticiones; ver el header Retry-After.
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ErrorInterno:
      description: 500 — Error interno; no expone trazas ni mensajes internos. Reintentar es seguro con idempotencia.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'